requirements.txt
flask
flask-cors
python-docx
requests
openai
python-dotenv
